refactor: reuse the kiosk return page for normal users
Instead of adding a separate return action to the loan detail page,
expose the existing return page to regular users. The page now scopes
its loan list to the current user's own loans when they are not
ADMIN/KIOSK, the view auth gate accepts any signed-in user, and the
"Palauta" nav link is shown to everyone. The shared dialog component
is dropped since the kiosk page already contains the return UI.
